+++ src/write.ts
@@ -1,7 +1,32 @@
 // Functions for converting properties to strings and to file system
 // TODO: Move all code here
 
+import { writeFileSync } from "fs";
+import Path from 'path';
+
 import { Settings, JsDoc } from "./types";
+
+/**
+ * Write index.ts file
+ *
+ * @param settings - Settings Object
+ * @param fileNamesToExport - List of file names that will be added to the index.ts file
+ */
+ export function writeIndexFile(settings: Settings, fileNamesToExport: string[]): void {
+  if (fileNamesToExport.length === 0) {
+    // Don't write an index file if its going to export nothing
+    return;
+  }
+  const exportLines = fileNamesToExport.map(fileName => `export * from './${fileName.replace(/\\/g, '/')}';`);
+  const fileContent = `${settings.fileHeader}\n\n${exportLines.join('\n').concat('\n')}`;
+  writeFileSync(Path.join(settings.typeOutputDirectory, 'index.ts'), fileContent);
+}
+
+export function getTypeFileNameFromSchema(schemaFileName: string, settings: Settings): string {
+  return schemaFileName.endsWith(`${settings.schemaFileSuffix}.ts`)
+    ? schemaFileName.substring(0, schemaFileName.length - `${settings.schemaFileSuffix}.ts`.length)
+    : schemaFileName.replace('.ts', '');
+}
 
 /**
  * Get all indent characters for this indent level
